design(FE): ReadParentPost 컴포넌트 제작 #163
